Personnummer (Personnummer v3.0.1)

Validate Swedish personal identity numbers Personnummer.

Functions

format(pnr)

Formats a personal identity number in short format.

Examples

iex> {_, p} = Personnummer.new("199001011234")
iex> Personnummer.format(p)
"900101-1234"
iex> {_, p} = Personnummer.new("199001010001")
iex> Personnummer.format(p)
"900101-0001"
iex> {_, p} = Personnummer.new("199001610001")
iex> Personnummer.format(p)
"900161-0001"

format(pnr, bool)

Formats a personal identity number in long format.

Examples

iex> {_, p} = Personnummer.new("9001011234")
iex> Personnummer.format(p, true)
"19900101-1234"
iex> Personnummer.format(p, false)
"900101-1234"

get_age(pnr)

Get the age of the person holding the personal identity number.

Examples

iex> now = DateTime.utc_now()
iex> {_, x} = Date.new(now.year - 20, now.month, now.day)
iex> pnr = "#{x.year}0101-1234"
iex> {_, p} = Personnummer.new(pnr)
iex> Personnummer.get_age(p)
20

is_coordination_number(pnr)

Returns true if the parsed personal identity number is a coordination number.

Examples

iex> {_, p} = Personnummer.new("800161-3294")
iex> Personnummer.is_coordination_number(p)
true

is_female?(pnr)

Returns true if the person behind the personal identity number is a female.

Examples

iex> {_, p} = Personnummer.new("19090903-6600")
iex> Personnummer.is_female?(p)
true

is_male?(pnr)

Returns true if the person behind the personal identity number is a male.

Examples

iex> {_, p} = Personnummer.new("19900101-0017")
iex> Personnummer.is_male?(p)
true

luhn(digits)

Calculate luhn checksum according to spec https://en.wikipedia.org/wiki/Luhn_algorithm.

Examples

iex> Personnummer.luhn("900101001")
7

new(pnr_string)

Construct a new Personnummer struct.

Examples

iex> Personnummer.new("19900101-0017")
{:ok,
 %Personnummer{
   control: 7,
   coordination: false,
   date: ~D[1990-01-01],
   separator: "-",
   serial: 1
 }}

valid?(_)

Checks if the personal identity number is valid. Requres a valid date and a valid last four digits.

Examples (for Persomnnummer type)

iex> p = %Personnummer{}
iex> Personnummer.valid?(p)
false
iex> {_, p} = Personnummer.new("19900101-0017")
iex> Personnummer.valid?(p)
true
iex> {_, p} = Personnummer.new("19900101-0018")
iex> Personnummer.valid?(p)
false

Examples (for string)

iex> Personnummer.valid?("19900101-0017")
true
iex> Personnummer.valid?("19900101-0019")
false
iex> Personnummer.valid?("bogus")
false
iex> Personnummer.valid?("903030-0017")
false